Revert "refactored git structure and some code, and small fixes"
