--- Main.py
@@ -1,9 +1,0 @@
-from FileReader import read_file
-
-
-def main():
-    read_file(input("Enter file name: "))
-
-
-if __name__ == "__main__":
-    main()
