+++ project_01.ged
@@ -1,0 +1,366 @@
+0 NOTE PREET DABHI 
+0 NOTE https://github.com/p-deathstroke/CS-555-Project.git
+0 HEAD
+1 SOUR Family Echo
+2 WWW http://www.familyecho.com/
+1 FILE My Family
+1 DATE 20 SEP 2021
+1 DEST ANSTFILE
+1 GEDC
+2 VERS 5.5.1
+2 FORM LINEAGE-LINKED
+1 SUBM @I1@
+2 NAME Preet
+1 SUBN
+1 CHAR UTF-8
+0 @I1@ INDI
+1 NAME Bruce /Wayne/
+2 GIVN Bruce
+2 SURN Wayne
+2 _MARNM Wayne
+1 SEX M
+1 BIRT
+2 DATE 17 APR 1978
+2 PLAC Gotham
+1 OCCU Batman
+2 PLAC Wayne Tech
+1 NOTE Interests: protecting gotham
+1 FAMS @F1@
+1 FAMS @F2@
+1 FAMS @F8@
+1 FAMC @F3@
+0 @I2@ INDI
+1 NAME Thomas /Wayne/
+2 GIVN Thomas
+2 SURN Wayne
+2 _MARNM Wayne
+1 SEX M
+1 BIRT
+2 DATE 2 MAR 1942
+1 DEAT Y
+2 DATE 26 JUL 1992
+1 FAMS @F3@
+0 @I3@ INDI
+1 NAME Martha /Wayne/
+2 GIVN Martha
+2 SURN Wayne
+2 _MARNM Wayne
+1 SEX F
+1 BIRT
+2 DATE 4 APR 1953
+1 DEAT Y
+2 DATE 26 JUL 1992
+1 FAMS @F3@
+0 @I4@ INDI
+1 NAME Talia /Al Ghul/
+2 GIVN Talia
+2 SURN Al Ghul
+2 _MARNM Al Ghul
+1 SEX F
+1 BIRT
+2 DATE 10 JUN 1984
+1 FAMS @F2@
+1 FAMC @F4@
+0 @I5@ INDI
+1 NAME Damian /Wayne/
+2 GIVN Damian
+2 SURN Wayne
+2 _MARNM Wayne
+1 SEX M
+1 BIRT
+2 DATE 7 AUG 2006
+1 FAMS @F5@
+1 FAMC @F2@
+0 @I6@ INDI
+1 NAME DIck /Grayson/
+2 GIVN DIck
+2 SURN Grayson
+2 _MARNM Grayson
+1 SEX M
+1 BIRT
+2 DATE 29 MAY 2000
+1 FAMS @F6@
+1 FAMS @F7@
+1 FAMC @F8@
+2 PEDI adopted
+1 FAMC @F9@
+2 PEDI birth
+0 @I7@ INDI
+1 NAME Barbra /Gordon/
+2 GIVN Barbra
+2 SURN Gordon
+2 _MARNM Gordon
+1 SEX F
+1 BIRT
+2 DATE 4 JUL 1985
+1 FAMS @F1@
+1 FAMC @F10@
+0 @I8@ INDI
+1 NAME Stephanie /Wayne/
+2 GIVN Stephanie
+2 SURN Wayne
+2 _MARNM Wayne
+1 SEX F
+1 BIRT
+2 DATE 22 SEP 2008
+1 FAMS @F11@
+1 FAMS @F12@
+1 FAMC @F1@
+0 @I9@ INDI
+1 NAME Ra's /Al Ghul/
+2 GIVN Ra's
+2 SURN Al Ghul
+2 _MARNM Al Ghul
+1 SEX M
+1 FAMS @F4@
+0 @I10@ INDI
+1 NAME Melisande /Al Ghul/
+2 GIVN Melisande
+2 SURN Al Ghul
+2 _MARNM Al Ghul
+1 SEX F
+1 FAMS @F4@
+0 @I11@ INDI
+1 NAME Ibn Al /Xu'ffash/
+2 GIVN Ibn Al
+2 SURN Xu'ffash
+2 _MARNM Xu'ffash
+1 SEX M
+1 BIRT
+2 DATE 26 OCT 2007
+1 FAMC @F2@
+0 @I12@ INDI
+1 NAME Lian /Wayne/
+2 GIVN Lian
+2 SURN Wayne
+2 _MARNM Wayne
+1 SEX F
+1 BIRT
+2 DATE 10 DEC 2008
+1 FAMS @F5@
+0 @I13@ INDI
+1 NAME Ranya /Wayne/
+2 GIVN Ranya
+2 SURN Wayne
+2 _MARNM Wayne
+1 SEX F
+1 BIRT
+2 DATE 12 APR 2012
+1 FAMS @F13@
+1 FAMC @F5@
+0 @I14@ INDI
+1 NAME Koriander /Star/
+2 GIVN Koriander
+2 SURN Star
+2 _MARNM Star
+1 SEX F
+1 BIRT
+2 DATE 8 JUL 2002
+1 FAMS @F7@
+0 @I15@ INDI
+1 NAME Barbra /Grayson/
+2 GIVN Barbra
+2 SURN Grayson
+2 _MARNM Grayson
+1 SEX F
+1 BIRT
+2 DATE 3 MAY 2001
+1 FAMS @F6@
+0 @I16@ INDI
+1 NAME Mary /Grayson/
+2 GIVN Mary
+2 SURN Grayson
+2 _MARNM Grayson
+1 SEX F
+1 BIRT
+2 DATE 3 FEB 2021
+1 FAMC @F6@
+0 @I17@ INDI
+1 NAME James /Grayson/
+2 GIVN James
+2 SURN Grayson
+2 _MARNM Grayson
+1 SEX M
+1 BIRT
+2 DATE 3 FEB 2021
+1 FAMC @F6@
+0 @I18@ INDI
+1 NAME John 2 /Grayson/
+2 GIVN John 2
+2 SURN Grayson
+2 _MARNM Grayson
+1 SEX M
+1 BIRT
+2 DATE 3 JUN 2022
+1 FAMS @F13@
+1 FAMC @F6@
+0 @I19@ INDI
+1 NAME Tim /Drake/
+2 GIVN Tim
+2 SURN Drake
+2 _MARNM Drake
+1 SEX M
+1 BIRT
+2 DATE 2 JAN 2005
+1 FAMC @F8@
+2 PEDI foster
+1 FAMC @F14@
+2 PEDI birth
+0 @I20@ INDI
+1 NAME Jason /Todd/
+2 GIVN Jason
+2 SURN Todd
+2 _MARNM Todd
+1 SEX M
+1 BIRT
+2 DATE 20 SEP 2022
+1 FAMC @F8@
+2 PEDI foster
+0 @I21@ INDI
+1 NAME Mary /Grayson/
+2 GIVN Mary
+2 SURN Grayson
+2 _MARNM Grayson
+1 SEX F
+1 BIRT
+2 DATE 7 NOV 1964
+1 DEAT Y
+2 DATE 11 MAR 2006
+1 FAMS @F9@
+0 @I22@ INDI
+1 NAME John /Grayson/
+2 GIVN John
+2 SURN Grayson
+2 _MARNM Grayson
+1 SEX M
+1 BIRT
+2 DATE 22 SEP 1960
+1 DEAT Y
+2 DATE 11 MAR 2006
+1 FAMS @F9@
+0 @I23@ INDI
+1 NAME Janet /Drake/
+2 GIVN Janet
+2 SURN Drake
+2 _MARNM Drake
+1 SEX M
+1 BIRT
+2 DATE 3 SEP 2010
+1 FAMS @F14@
+0 @I24@ INDI
+1 NAME Jack /Drake/
+2 GIVN Jack
+2 SURN Drake
+2 _MARNM Drake
+1 SEX M
+1 BIRT
+2 DATE 5 OCT 1975
+1 DEAT Y
+2 DATE 3 SEP 2010
+1 FAMS @F14@
+0 @I25@ INDI
+1 NAME Barbra-Kean /Gordon/
+2 GIVN Barbra-Kean
+2 SURN Gordon
+2 _MARNM Gordon
+1 SEX F
+1 BIRT
+2 DATE 10 AUG 1954
+1 FAMS @F10@
+0 @I26@ INDI
+1 NAME James /Gordon/
+2 GIVN James
+2 SURN Gordon
+2 _MARNM Gordon
+1 SEX F
+1 BIRT
+2 DATE 4 JUL 1950
+1 FAMS @F10@
+0 @I27@ INDI
+1 NAME Gar /Grayson/
+2 GIVN Gar
+2 SURN Grayson
+2 _MARNM Grayson
+1 SEX M
+1 BIRT
+2 DATE 11 AUG 2032
+1 FAMC @F13@
+0 @F1@ FAM
+1 HUSB @I1@
+1 WIFE @I7@
+1 CHIL @I8@
+1 _CURRENT Y
+0 @F2@ FAM
+1 HUSB @I1@
+1 WIFE @I4@
+1 CHIL @I5@
+1 CHIL @I11@
+1 MARR
+2 DATE 6 FEB 1995
+2 PLAC Gotham
+1 DIV
+2 DATE 7 JUN 2007
+1 _CURRENT N
+0 @F3@ FAM
+1 HUSB @I2@
+1 WIFE @I3@
+1 CHIL @I1@
+1 _CURRENT Y
+0 @F4@ FAM
+1 HUSB @I9@
+1 WIFE @I10@
+1 CHIL @I4@
+1 _CURRENT Y
+0 @F5@ FAM
+1 HUSB @I5@
+1 WIFE @I12@
+1 CHIL @I13@
+1 _CURRENT Y
+0 @F6@ FAM
+1 HUSB @I6@
+1 WIFE @I15@
+1 CHIL @I16@
+1 CHIL @I17@
+1 CHIL @I18@
+1 _CURRENT Y
+0 @F7@ FAM
+1 HUSB @I6@
+1 WIFE @I14@
+1 MARR
+2 DATE 23 MAY 2018
+2 PLAC Gotham
+1 DIV
+2 DATE 9 FEB 2020
+1 _CURRENT N
+0 @F8@ FAM
+1 HUSB @I1@
+1 CHIL @I6@
+1 CHIL @I19@
+1 CHIL @I20@
+0 @F9@ FAM
+1 HUSB @I22@
+1 WIFE @I21@
+1 CHIL @I6@
+1 _CURRENT Y
+0 @F10@ FAM
+1 HUSB @I25@
+1 WIFE @I26@
+1 CHIL @I7@
+1 _CURRENT Y
+0 @F11@ FAM
+1 WIFE @I8@
+1 _CURRENT N
+0 @F12@ FAM
+1 WIFE @I8@
+1 _CURRENT N
+0 @F13@ FAM
+1 HUSB @I18@
+1 WIFE @I13@
+1 CHIL @I27@
+1 _CURRENT Y
+0 @F14@ FAM
+1 HUSB @I24@
+1 WIFE @I23@
+1 CHIL @I19@
+1 _CURRENT Y
+0 TRLR
